+++ src/components/NodeImpl.cpp
@@ -6,7 +6,7 @@
 
 #include <utility>
 
-EndSystem::EndSystem(std::string name, int dpr): Node(std::move(name), END_SYSTEM, dpr) {
+EndSystem::EndSystem(std::string name, uint32_t dpr): Node(std::move(name), END_SYSTEM, dpr) {
     Port _port;
     this->port = _port;
 }
@@ -23,21 +23,21 @@ std::string &EndSystem::toString() {
     return this->name;
 }
 
-Switch::Switch(std::string name, int dpr): Node(std::move(name), SWITCH, dpr) {
+Switch::Switch(std::string name, uint32_t dpr): Node(std::move(name), SWITCH, dpr) {
     for (int i = 0; i < portNum; ++i) {
         Port port;
         ports.push_back(port);
     }
 }
 
-Switch::Switch(std::string name, int dpr, int portNum) : Node(std::move(name), SWITCH, dpr), portNum(portNum) {
+Switch::Switch(std::string name, uint32_t dpr, uint8_t portNum) : Node(std::move(name), SWITCH, dpr), portNum(portNum) {
     for (int i = 0; i < portNum; ++i) {
         Port port;
         ports.push_back(port);
     }
 }
 
-int Switch::getPortNum() const {
+uint8_t Switch::getPortNum() const {
     return portNum;
 }
 
@@ -83,11 +83,11 @@ void GraphNode::setState(NODE_SEARCH_STATE _state) {
     GraphNode::state = _state;
 }
 
-int GraphNode::getHop() const {
+uint8_t GraphNode::getHop() const {
     return hop;
 }
 
-void GraphNode::setHop(int _hop) {
+void GraphNode::setHop(uint8_t _hop) {
     GraphNode::hop = _hop;
 }
